gofmt, more script-local package ids consts, fix bug saving script cached ints

+++ pkg/server/db/player.go
@@ -397,7 +397,7 @@ func SavePlayer(player *world.Player) {
 		var val string
 		switch value.(type) {
 		case int64:
-			val = "i" + strconv.FormatInt(int64(value.(int)), 10)
+			val = "i" + strconv.FormatInt(value.(int64), 10)
 		case int:
 			val = "i" + strconv.FormatInt(int64(value.(int)), 10)
 		case uint:
